feat: add AppContext;use Animated components and useAnimatedStyle hooks;
The app context was needed to use the popup's state and theme shared values inside animation worklets. Added portal to render the children in root of the dom tree where the PortalProvider exists. The item list has 2 copies, one is hidden and on top of the real item list. The hidden item is enclosed in a Portal Component (ListAnimatedContent.tsx). As soon as long press is observed, we animate the hidden item list to be visible and translates it's Y position and the real one is hidden with making it's opacity to 0. The shared context value of Popup's state is observed by blur view which is in the bottom of the screen and hidden, as soon as long press is observed, we change it's opacity to 1 and change it's top style from WINDOW_HEIGHT to 0 so that it's visible